Use actual DDK headers in wrapped driver code
