+++ usersim.sln
@@ -11,6 +11,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "sample", "sample\sample.vcx
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "tests", "tests\tests.vcxproj", "{C7B63CA3-A0B8-4EE3-85C8-A17CB40440BF}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "usersim_dll_skeleton", "usersim_dll_skeleton\usersim_dll_skeleton.vcxproj", "{1937DB41-F3EB-4955-A636-6386DCB394F6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|ARM64 = Debug|ARM64
@@ -131,6 +133,30 @@ Global
 		{C7B63CA3-A0B8-4EE3-85C8-A17CB40440BF}.RelWithDebInfo|x64.Build.0 = Release|x64
 		{C7B63CA3-A0B8-4EE3-85C8-A17CB40440BF}.RelWithDebInfo|x86.ActiveCfg = Release|Win32
 		{C7B63CA3-A0B8-4EE3-85C8-A17CB40440BF}.RelWithDebInfo|x86.Build.0 = Release|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Debug|ARM64.ActiveCfg = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Debug|ARM64.Build.0 = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Debug|x64.ActiveCfg = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Debug|x64.Build.0 = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Debug|x86.ActiveCfg = Debug|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Debug|x86.Build.0 = Debug|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.MinSizeRel|ARM64.ActiveCfg = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.MinSizeRel|ARM64.Build.0 = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.MinSizeRel|x64.ActiveCfg = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.MinSizeRel|x64.Build.0 = Debug|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.MinSizeRel|x86.ActiveCfg = Debug|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.MinSizeRel|x86.Build.0 = Debug|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Release|ARM64.ActiveCfg = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Release|ARM64.Build.0 = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Release|x64.ActiveCfg = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Release|x64.Build.0 = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Release|x86.ActiveCfg = Release|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.Release|x86.Build.0 = Release|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.RelWithDebInfo|ARM64.ActiveCfg = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.RelWithDebInfo|ARM64.Build.0 = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.RelWithDebInfo|x64.ActiveCfg = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.RelWithDebInfo|x64.Build.0 = Release|x64
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.RelWithDebInfo|x86.ActiveCfg = Release|Win32
+		{1937DB41-F3EB-4955-A636-6386DCB394F6}.RelWithDebInfo|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
