+++ usersim_dll_skeleton/dll_skeleton.h
@@ -1,0 +1,67 @@
+// Copyright (c) Microsoft Corporation
+// SPDX-License-Identifier: MIT
+#pragma once
+#include <sal.h>
+#include <winsock2.h>
+#include <windows.h>
+
+typedef struct _wdfdriver WDFDRIVER;
+
+typedef struct _driver_object DRIVER_OBJECT, *PDRIVER_OBJECT;
+
+typedef struct _UNICODE_STRING
+{
+    USHORT Length;
+    USHORT MaximumLength;
+    _Field_size_bytes_part_opt_(MaximumLength, Length) PWCH Buffer;
+} UNICODE_STRING, *PUNICODE_STRING;
+typedef const UNICODE_STRING* PCUNICODE_STRING;
+
+typedef NTSTATUS(DRIVER_INITIALIZE)(_In_ PDRIVER_OBJECT driver_object, _In_ PUNICODE_STRING registry_path);
+
+typedef struct _WDFDEVICE_INIT WDFDEVICE_INIT, *PWDFDEVICE_INIT;
+
+typedef NTSTATUS(EVT_WDF_DRIVER_DEVICE_ADD)(_In_ WDFDRIVER driver, _Inout_ PWDFDEVICE_INIT device_init);
+typedef EVT_WDF_DRIVER_DEVICE_ADD* PFN_WDF_DRIVER_DEVICE_ADD;
+
+typedef void(EVT_WDF_DRIVER_UNLOAD)(_In_ WDFDRIVER driver);
+typedef EVT_WDF_DRIVER_UNLOAD* PFN_WDF_DRIVER_UNLOAD;
+
+typedef struct _WDF_DRIVER_CONFIG
+{
+    ULONG Size;
+    PFN_WDF_DRIVER_DEVICE_ADD EvtDriverDeviceAdd;
+    PFN_WDF_DRIVER_UNLOAD EvtDriverUnload;
+    ULONG DriverInitFlags;
+    ULONG DriverPoolTag;
+} WDF_DRIVER_CONFIG, *PWDF_DRIVER_CONFIG;
+
+typedef struct _wdfdriver
+{
+    WDF_DRIVER_CONFIG config;
+} WDFDRIVER;
+
+#define WDF_DRIVER_GLOBALS_NAME_LEN (32)
+
+typedef struct _WDF_DRIVER_GLOBALS
+{
+    // backpointer to the handle for this driver
+    WDFDRIVER Driver;
+
+    // Flags indicated by the driver during create
+    ULONG DriverFlags;
+
+    // Tag generated by WDF for the driver.  Tag used by allocations made on
+    // behalf of the driver by WDF.
+    ULONG DriverTag;
+
+    CHAR DriverName[WDF_DRIVER_GLOBALS_NAME_LEN];
+
+    // If TRUE, the stub code will capture DriverObject->DriverUnload and insert
+    // itself first in the unload chain.  If FALSE, DriverUnload is left alone
+    // (but WDF will not be notified of unload and there will be no auto cleanup).
+    BOOLEAN DisplaceDriverUnload;
+
+} WDF_DRIVER_GLOBALS, *PWDF_DRIVER_GLOBALS;
+
+typedef VOID (*WDFFUNC)(VOID);
